feat(dedicated.nasha): refacto details page
ref:MANAGER-10906
