+++ packages/manager/apps/nasha-react/src/pages/details/[serviceName]/suscriptions.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { fetchNashaServiceInfos } from '../../../api/nasha-react';
+
+function Subscriptions(props: { serviceName: string }) {
+  const { serviceName } = props;
+  const { isLoading, isError, data } = useQuery(
+    ['serviceInfos', { serviceName }],
+    fetchNashaServiceInfos,
+  );
+
+  if (isLoading) {
+    return <span>Loading...</span>;
+  }
+
+  if (isError) {
+    return <span>Error...</span>;
+  }
+
+  const count = data?.length;
+  if (count === 0) {
+    return <></>;
+  }
+  return (
+    <>
+      <ul>{JSON.stringify(data)}</ul>
+      <ul>
+        <button>Configurer le renouvellement</button>
+        <button>Anticiper le paiement</button>
+        <button>Configurer le renouvellement</button>
+      </ul>
+    </>
+  );
+}
+
+export default Subscriptions;
